requirements.txt
streamlit>=1.30.0
google-genai>=0.4.0
Pillow>=10.0.0
ezdxf>=1.4.0
pydantic>=2.5.0
python-dotenv>=1.0.0
opencv-python>=4.8.0

# Accuracy enhancements
paddlepaddle>=2.5.0
paddleocr>=2.7.0
rapidfuzz>=3.0.0
